// src/app/shared/types/api-types/bire-item-attribute-dto.interface.ts
export interface BireItemAttributeDTO {
  familyCode?: string;
  structureType?: string;
  variantCode?: string;
  chapter?: string;
  sheet?: string;
  marks?: string;
  section?: string;
  subject?: string;
  attributeId?: string;
  attributeValue?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}